src/requirements.txt
Django
djangorestframework
drf-spectacular
Pillow
psycopg2-binary
celery
redis
django-redis
gunicorn
python-dotenv
